compiling partial implementation of domino free biclique cover

+++ inc/Cigar.hpp
@@ -71,7 +71,7 @@ public:
     Alignment(const string& s);
 
     // Return the sequence lengths of the {query,ref} in a pair
-    void compute_lengths(pair<uint64_t,uint64_t>& lengths);
+    void compute_lengths(pair<size_t,size_t>& lengths);
 
     // Given a pair of start indexes {query_start,ref_start} walk through the alignment, updating the indexes in the
     // pair to retrace the coordinates of the path in the alignment matrix that produces this cigar. Return true as long
